backend-ia/requirements.txt
ultralytics>=8.0.0
flask>=3.0.0
flask-cors>=4.0.0
pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.4.0
paho-mqtt>=2.0.0
pymongo>=4.6.0
anthropic>=0.25.0
python-dotenv>=1.0.0
requests>=2.31.0
